+++ src/test/java/nl/cerios/reactive/chocolate/factory/model/PeanutTest.kt
@@ -7,13 +7,13 @@ internal class PeanutTest {
 
   @Test
   fun json() {
-    val quality = 0.25F
+    val quality = 0.25
 
     val peanut1 = Peanut(quality)
     val json = peanut1.toJson()
     val peanut2 = Peanut.fromJson(json)
 
-    assertEquals(quality, json.getFloat("quality"))
+    assertEquals(quality, json.getDouble("quality"))
     assertEquals("""{"quality":0.25}""", json.encode())
     assertEquals(peanut1, peanut2)
   }
